default/create: 특강 선택 API 단위 테스트 추가
