+++ default-todo.md
@@ -1,0 +1,40 @@
+## Default
+- [x] 아키텍처 준수를 위한 애플리케이션 패키지 설계
+- [ ] 특강 도메인 테이블 설계 및 목록/신청 등 기본 기능 구현
+- [ ] 각 기능에 대한 **단위 테스트** 작성
+
+---
+### 세부 API
+- [ ] 특강 선택 API (=신청 가능 강의 목록 조회 API)
+  * 날짜별로 현재 신청 가능한 특강 목록을 조회하는 API 를 작성합니다.
+  * 특강의 정원은 30명으로 고정이며, 사용자는 각 특강에 신청하기전 목록을 조회해볼 수 있어야 합니다.
+
+  **단위 테스트**
+  - [ ] 파라미터 validator
+  - [x] 요청된 날짜에 해당되는 강의가 없으면 예외발생
+  - [x] 요청된 날짜에 해당되는 강의 목록 조회
+
+  > (요청한 날짜에 있는 강의 + "신청가능 인원수 = 0초과") 강의 목록 조회
+
+<br>
+
+- [ ] 특강 신청 완료 목록 조회 API
+  * 특정 userId 로 신청 완료된 특강 목록을 조회하는 API 를 작성합니다.
+  * 각 항목은 특강 ID 및 이름, 강연자 정보를 담고 있어야 합니다.
+
+  **단위 테스트**
+  - [ ] 파라미터 validator
+
+<br>
+
+- [ ] 특강 신청 API
+  * 특정 userId 로 선착순으로 제공되는 특강을 신청하는 API 를 작성합니다.
+  * 동일한 신청자는 동일한 강의에 대해서 한 번의 수강 신청만 성공할 수 있습니다.
+  * 특강은 선착순 30명만 신청 가능합니다.
+  * 이미 신청자가 30명이 초과되면 이후 신청자는 요청을 실패합니다.
+
+  **단위 테스트**
+  - [ ] 파라미터 validator
+  - [ ] 중복 신청 예외발생 (동일한 신청자는 동일한 강의에 대해서 한 번의 수강 신청만 가능)
+  - [ ] 남은 자리가 없으면 예외발생 (30명이 초과된 상태)
+  - [ ] (신청 가능한 자리수 - 1)이 정상 동작하는지 확인
